Mode BSMPSim to bsmp subpackage
